==> R/01_misc_ABS_data_prep.Rmd <==
---
title: "Misc ABS data preparation"
# subtitle: 
# author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    # css: custom.css
    toc: true
    toc_depth: 4
    toc_float: true
    number_sections: true
    theme: united 
    highlight: pygments 
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, 
                      fig.width=9, fig.height=7, dpi=300, out.width="900px", out.height="700px")

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

options(scipen=999)
set.seed(12345)

library(pacman) 
p_load(tidyverse, readr, readxl, lubridate, 
       scales, janitor, kableExtra, 
       sjmisc,
       sf, tmap)

tmap_mode("view")
```

```{r geodata_prep, eval=FALSE}
p_load(rmapshaper)

# SA3 outlines
SA3 <- st_read("./data/geo/1270055001_sa3_2016_aust_shape/SA3_2016_AUST.shp", 
               stringsAsFactors = FALSE)
# View(SA3)
# st_crs(SA3)
saveRDS(SA3, file = "./data/geo/SA3_2016_AUST.rds")

# other teritories and zeroes
SA3 %>% 
  filter(SA4_CODE16 == 901)

# Keeping Jervis bay in?
# removing Lord Howe Island?
SA3 %>% 
  filter(SA3_CODE16 %in% c(90101, 90102, 90104, 10803))

SA3 %>% 
  filter(AREASQKM16 == 0)  

SA3 %<>% 
  # filter(SA4_CODE16 != 901) %>% 
  filter(! SA3_CODE16 %in% c(90101, 90102, 90104, 10803)) %>% 
  filter(AREASQKM16 != 0) %>% 
  st_transform(3112) 

nrow(SA3)

SA3 <- ms_simplify(SA3, keep = 0.05, weighting = 0.7) # default settings
# SA3 <- ms_simplify(SA3, keep = 0.04, weighting = 0.8) # bit more aggressive

nrow(SA3)
# qtm(SA3)

saveRDS(SA3, file = "./data/geo/SA3_2016_AUST_clean.rds")

# Mesh blocks of 'Blue Mountains - South' SA3 have 0 pops
# replaced with 'normal' centroid

SA3_centr <- read_csv("data/geo/SA3_2016_pop_center_median.zip") %>% 
  select(-OBJECTID) %>% 
  rename(lon = XCoord, 
         lat = YCoord) %>% 
  mutate(SA3_CODE16 = as.character(SA3_CODE16)) %>% 
  right_join(select(st_drop_geometry(SA3), SA3_CODE16, SA3_NAME16)) %>% 
  filter(! SA3_CODE16 %in% c(90101, 90102, 90104, 10803)) %>% 
  mutate(lon = ifelse(SA3_CODE16 == "12402", 150.287952, lon)) %>% 
  mutate(lat = ifelse(SA3_CODE16 == "12402", -33.993425, lat)) %>% 
  st_as_sf(coords = c("lon", "lat"), crs = 4326, remove = FALSE) %>% 
  st_transform(3112) %>% 
  select(SA3_CODE16, everything())

# qtm(SA3_centr)

saveRDS(SA3_centr, file = "./data/geo/SA3_centr.rds")

p_unload(rmapshaper)
```

```{r include=FALSE}
SA3 <- readRDS(file = "./data/geo/SA3_2016_AUST.rds")
# SA3_centr <- readRDS(file = "./data/geo/SA3_centr.rds")

pop_sa1 <- read_csv("data/ABS/2016 Census GCP Statistical Area 1 for AUST/2016Census_G01_AUS_SA1.csv.gz") %>% 
  select(SA1_7DIGITCODE_2016, Tot_P_P) %>% 
  rename(SA1_7DIG16 = SA1_7DIGITCODE_2016,
         pop_16 = Tot_P_P)

SA1 <- st_read("./data/geo/1270055001_sa1_2016_aust_shape/SA1_2016_AUST.shp", 
               stringsAsFactors = FALSE) %>% 
  st_drop_geometry() %>% 
  select(SA1_MAIN16, SA3_CODE16, SA3_NAME16) %>% 
  mutate(SA1_MAIN16 = as.numeric(SA1_MAIN16))
```

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

# Population present

```{r}
pop_present <- readxl::read_xls("./Data/ABS/Present Population OAD and Natural Increase.xls", 
                                sheet = "Daily Population Present", 
                                skip = 1, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  dplyr::select(-x1, -x9, -natural_increase, -days_in_quarter, -daily_change_in_erp) %>% 
  dplyr::rename(date = x2) %>% 
  dplyr::mutate(date = ymd(date))

saveRDS(pop_present, "./Data/ABS/pop_present.Rds")

pop_present_month <- pop_present %>% 
  filter(date != ymd("2017-12-31")) %>% 
  mutate(start = floor_date(date, unit = "month"),
         end = ceiling_date(date, unit = "month", change_on_boundary = TRUE) - days(1)) %>% 
  group_by(start, end) %>% 
  summarize(mean_month_pop_present = mean(daily_physically_present)) %>% 
  ungroup()

saveRDS(pop_present_month, "./Data/ABS/pop_present_month.Rds")
```

```{r}
ggplot() +
  geom_line(data = pop_present,
            aes(x=date, y=daily_physically_present), 
            size = 1, colour = "darkorchid4") +
  geom_segment(data = pop_present_month,
               aes(x=start, xend=end, y=mean_month_pop_present, yend=mean_month_pop_present)) +
  scale_x_date(date_breaks = "2 months", date_minor_breaks = "1 month") + 
  theme_minimal() + xlab("") + ylab("Daily physically present")
```


<!-- ------------------------------------------------------------ --> 

# ERP 18

```{r eval=FALSE, include=FALSE}
NSW <- readxl::read_excel("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/raw/32180ds0001_2017-18.xls", 
                          sheet = "Table 1", skip = 7, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  filter(!is.na(sa2_code)) %>% 
  dplyr::select(-km2, -persons_km2) %>% 
  dplyr::rename(ERP_17 = "no_11", ERP_18 = "no_12", 
                ERP_ch_abs = "no_14", ERP_ch_per = "percent", 
                NI = "no_17", NIM = "no_18", NOM = "no_19")

# View(NSW)

Vic <- readxl::read_excel("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/raw/32180ds0001_2017-18.xls", 
                          sheet = "Table 2", skip = 7, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  filter(!is.na(sa2_code)) %>% 
  dplyr::select(-km2, -persons_km2) %>% 
  dplyr::rename(ERP_17 = "no_11", ERP_18 = "no_12", 
                ERP_ch_abs = "no_14", ERP_ch_per = "percent", 
                NI = "no_17", NIM = "no_18", NOM = "no_19")

# View(Vic)

Qld <- readxl::read_excel("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/raw/32180ds0001_2017-18.xls", 
                          sheet = "Table 3", skip = 7, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  filter(!is.na(sa2_code)) %>% 
  dplyr::select(-km2, -persons_km2) %>% 
  dplyr::rename(ERP_17 = "no_11", ERP_18 = "no_12", 
                ERP_ch_abs = "no_14", ERP_ch_per = "percent", 
                NI = "no_17", NIM = "no_18", NOM = "no_19")

# View(Qld)

SA <- readxl::read_excel("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/raw/32180ds0001_2017-18.xls", 
                         sheet = "Table 4", skip = 7, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  filter(!is.na(sa2_code)) %>% 
  dplyr::select(-km2, -persons_km2) %>% 
  dplyr::rename(ERP_17 = "no_11", ERP_18 = "no_12", 
                ERP_ch_abs = "no_14", ERP_ch_per = "percent", 
                NI = "no_17", NIM = "no_18", NOM = "no_19")

# View(SA)

WA <- readxl::read_excel("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/raw/32180ds0001_2017-18.xls", 
                         sheet = "Table 5", skip = 7, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  filter(!is.na(sa2_code)) %>% 
  dplyr::select(-km2, -persons_km2) %>% 
  dplyr::rename(ERP_17 = "no_11", ERP_18 = "no_12", 
                ERP_ch_abs = "no_14", ERP_ch_per = "percent", 
                NI = "no_17", NIM = "no_18", NOM = "no_19")

# View(WA)

Tas <- readxl::read_excel("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/raw/32180ds0001_2017-18.xls", 
                          sheet = "Table 6", skip = 7, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  filter(!is.na(sa2_code)) %>% 
  dplyr::select(-km2, -persons_km2) %>% 
  dplyr::rename(ERP_17 = "no_11", ERP_18 = "no_12", 
                ERP_ch_abs = "no_14", ERP_ch_per = "percent", 
                NI = "no_17", NIM = "no_18", NOM = "no_19")

# View(Tas)

NT <- readxl::read_excel("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/raw/32180ds0001_2017-18.xls", 
                         sheet = "Table 7", skip = 7, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  filter(!is.na(sa2_code)) %>% 
  dplyr::select(-km2, -persons_km2) %>% 
  dplyr::rename(ERP_17 = "no_11", ERP_18 = "no_12", 
                ERP_ch_abs = "no_14", ERP_ch_per = "percent", 
                NI = "no_17", NIM = "no_18", NOM = "no_19")

# View(NT)

ACT <- readxl::read_excel("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/raw/32180ds0001_2017-18.xls", 
                          sheet = "Table 8", skip = 7, col_names = TRUE) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>%
  filter(!is.na(sa2_code)) %>% 
  dplyr::select(-km2, -persons_km2) %>% 
  dplyr::rename(ERP_17 = "no_11", ERP_18 = "no_12", 
                ERP_ch_abs = "no_14", ERP_ch_per = "percent", 
                NI = "no_17", NIM = "no_18", NOM = "no_19")

# View(ACT)

ERP_SA2_2018 <- bind_rows(NSW, Vic, Qld, SA, WA, Tas, NT, ACT) %>% 
  # mutate(SA2_NAME = gsub("`", "'", SA2_NAME)) %>% # rogue character for apostrophe, not matching ArcGIS!
  write_csv("./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/clean/ERP_SA2_2018.csv", na = "", append = FALSE)

saveRDS(ERP_SA2_2018, "./Data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/clean/ERP_SA2_2018.Rds")
rm(NSW, Vic, Qld, SA, WA, Tas, NT, ACT) 

# summary(ERP_SA2_2016$ERP)
```

Using file *'Released at 11.30am (Canberra time) 27 March 2019'*.

```{r}
ERP_SA3_2018 <- readRDS("./data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/clean/ERP_SA2_2018.Rds") %>% 
  group_by(s_t_name, sa3_code, sa3_name) %>% 
  summarise(ERP_18 = sum(ERP_18)) %>% 
  ungroup()
```

Spatial resolution - **`r nrow(ERP_SA3_2018)`** SA3s.   

```{r}
ggplot(ERP_SA3_2018, aes(x = s_t_name, y = ERP_18)) +
  geom_boxplot(varwidth = TRUE) + 
  xlab("") + theme_minimal()
```

<!-- ------------------------------------------------------------ --> 

# Census pop 

Data from https://datapacks.censusdata.abs.gov.au/datapacks/

## POA

```{r}
pop_POA <- read_csv("data/ABS/2016 Census GCP Postal Areas for AUST/2016Census_G04A_AUS_POA.csv.gz") %>% 
  left_join(read_csv("data/ABS/2016 Census GCP Postal Areas for AUST/2016Census_G04B_AUS_POA.csv.gz")) %>% 
  select(-ends_with("_F"), -ends_with("_M")) %>% 
  select(POA_CODE_2016, Age_yr_18_P, Age_yr_19_P, matches("(Age_yr_\\d\\d_\\d\\d_P)"), Age_yr_100_yr_over_P, Tot_P) %>% 
  select(-Age_yr_10_14_P, -Age_yr_15_19_P) %>% 
  mutate(pop_cen_18plus = rowSums(select(., Age_yr_18_P:Age_yr_100_yr_over_P), na.rm = TRUE)) %>% 
  mutate(POA_CODE_2016 = str_replace(POA_CODE_2016, "POA", "")) %>% 
  select(POA_CODE_2016, pop_cen_18plus) %>% 
  rename(POA_CODE16 = POA_CODE_2016) %>% 
  arrange(POA_CODE16)
```

Spatial resolution - **`r nrow(pop_POA)`** postcodes.   

```{r}
ggplot(pop_POA, aes(pop_cen_18plus)) + 
  geom_histogram() +
  ylab("Number of postcodes") + xlab("Population") + theme_minimal()
```

```{r eval=FALSE, include=FALSE}
POA_centr %>% 
  left_join(pop_POA) %>% 
  tm_shape() +
  tm_bubbles("pop_cen_18plus", scale=.5)
```

## SA3 

```{r}
pop_SA3 <- read_csv("data/ABS/2016 Census GCP Statistical Area 3 for AUST/2016Census_G04A_AUS_SA3.csv.gz") %>% 
  left_join(read_csv("data/ABS/2016 Census GCP Statistical Area 3 for AUST/2016Census_G04B_AUS_SA3.csv.gz")) %>% 
  select(-ends_with("_F"), -ends_with("_M")) %>% 
  select(SA3_CODE_2016, Age_yr_18_P, Age_yr_19_P, matches("(Age_yr_\\d\\d_\\d\\d_P)"), Age_yr_100_yr_over_P, Tot_P) %>% 
  select(-Age_yr_10_14_P, -Age_yr_15_19_P) %>% 
  mutate(pop_cen_18plus = rowSums(select(., Age_yr_18_P:Age_yr_100_yr_over_P), na.rm = TRUE)) %>% 
  select(SA3_CODE_2016, Tot_P, pop_cen_18plus) %>% 
  rename(SA3_CODE16 = SA3_CODE_2016,
         total_population = Tot_P) %>% 
  arrange(SA3_CODE16)
```

Spatial resolution - **`r nrow(pop_SA3)`** SA3s.   

```{r}
ggplot(pop_SA3, aes(pop_cen_18plus)) + 
  geom_histogram() +
  ylab("Number of SA3") + xlab("Population") + theme_minimal()
```

```{r eval=FALSE, include=FALSE}
POA_centr %>% 
  left_join(pop_SA3) %>% 
  tm_shape() +
  tm_bubbles("pop_cen_18plus", scale=.5)
```


<!-- ------------------------------------------------------------ --> 

# Remotness

Data are from *1270.0.55.005 - Australian Statistical Geography Standard (ASGS): Volume 5 - Remoteness Structure, July 2016* [info](https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/1270.0.55.005July%202016?OpenDocument).

`No usual address` and `Migratory - Offshore - Shipping` categories are excluded. 

Data are on SA1 level. Population of SA1 was used to find majority classification within each SA3.

```{r include=FALSE}
remote <- read_csv("data/ABS/1270055005_ra_2016_aust_csv/RA_2016_AUST.zip", 
                   col_types = cols(AREA_ALBERS_SQKM = col_skip(), 
                                    MB_CODE_2016 = col_skip(), RA_CODE_2016 = col_integer(), 
                                    STATE_CODE_2016 = col_skip(), STATE_NAME_2016 = col_skip())) %>% 
  # group_by(SA1_MAINCODE_2016, RA_NAME_2016) %>% 
  # mutate(n1 = n()) %>% 
  # ungroup() %>% 
  # group_by(SA1_MAINCODE_2016) %>% 
  # mutate(n2 = n()) %>% 
  # ungroup() %>% 
  # filter(n1 != n2)
  rename(SA1_MAIN16 = SA1_MAINCODE_2016,
         SA1_7DIG16 = SA1_7DIGITCODE_2016) %>% 
  group_by(SA1_MAIN16) %>% 
  filter(row_number()==1) %>% 
  ungroup() %>% 
  filter(!str_detect(RA_NAME_2016, "Migratory - Offshore - Shipping")) %>% 
  filter(!str_detect(RA_NAME_2016, "No usual address"))

remote %<>% 
  left_join(pop_sa1) %>% 
  left_join(SA1)

remote %<>% 
  rename(pop_sa1_16 = pop_16) %>% 
  group_by(SA3_CODE16) %>% 
  mutate(pop_sa3_16 = sum(pop_sa1_16)) %>% 
  ungroup() 
```

Original data

```{r}
remote %>% 
  frq(RA_NAME_2016)
```

SA3 assigned to majority

```{r}
remote_ra <- 
  remote %>% 
  group_by(SA3_CODE16, RA_CODE_2016) %>% 
  mutate(pop_sa3_16_cat = sum(pop_sa1_16)) %>% 
  ungroup() %>% 
  mutate(pop_sa3_16_s = pop_sa3_16_cat/pop_sa3_16) %>% 
  arrange(SA3_CODE16, desc(pop_sa3_16_s)) %>% 
  group_by(SA3_CODE16, RA_CODE_2016) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  select(-SA1_MAIN16, -SA1_7DIG16, -pop_sa1_16) %>% 
  select(SA3_CODE16, SA3_NAME16, everything())

remote_ra_one <- 
  remote_ra %>% 
  group_by(SA3_CODE16) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  filter(! SA3_CODE16 %in% c(90101, 90102, 90104, 10803))
```

Share of majority category

```{r}
remote_ra_one %>% 
  descr(pop_sa3_16_s)

ggplot(remote_ra_one, aes(pop_sa3_16_s)) + 
  geom_histogram() +
  ylab("Number of SA3") + xlab("Share of majority category") + theme_minimal()
```

Map 

```{r}
SA3 %>% 
  left_join(remote_ra_one) %>% 
  tm_shape() + 
  tm_polygons(col = "RA_NAME_2016", alpha = .5, border.col = NA, id = "SA3_NAME16")
```

```{r include=FALSE}
write_csv(remote_ra_one, "data/model_input/remote_sa3_18.csv")
```


<!-- ------------------------------------------------------------ --> 

# Level of urbanization

Data are from *1270.0.55.001 - Australian Statistical Geography Standard (ASGS): Volume 1 - Main Structure and Greater Capital City Statistical Areas, July 2016* [info](https://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/1270.0.55.001Main+Features10018July%202016?OpenDocument)

`No usual address` and `Migratory - Offshore - Shipping` categories are excluded. 

Data are on SA1 level. Population of SA1 was used to find majority classification within each SA3.


```{r include=FALSE}
urban <- read_csv("data/ABS/1270055004_sa1_ucl_sosr_sos_2016_aust_csv/raw/SA1_UCL_SOSR_SOS_2016_AUST.zip", 
                  col_types = cols(AREA_ALBERS_SQKM = col_skip(), 
                                   SOSR_CODE_2016 = col_integer(), 
                                   SOS_CODE_2016 = col_integer(), 
                                   STE_CODE_2016 = col_skip(), STE_NAME_2016 = col_skip())) %>% 
  rename(SA1_MAIN16 = SA1_MAINCODE_2016,
         SA1_7DIG16 = SA1_7DIGITCODE_2016) %>% 
  select(-starts_with("UCL")) %>% 
  filter(!str_detect(SOSR_NAME_2016, "Migratory - Offshore - Shipping")) %>% 
  filter(!str_detect(SOSR_NAME_2016, "No usual address"))

urban %<>% 
  left_join(pop_sa1) %>% 
  left_join(SA1)

urban %<>% 
  rename(pop_sa1_16 = pop_16) %>% 
  group_by(SA3_CODE16) %>% 
  mutate(pop_sa3_16 = sum(pop_sa1_16)) %>% 
  ungroup() 

rm(pop_sa1, SA1)
```

### Section of State Range (SOSR) 

Original data

```{r}
urban %>% 
  frq(SOSR_NAME_2016)
```

SA3 assigned to majority

```{r}
urban_sosr <- 
  urban %>% 
  select(-SOS_CODE_2016, -SOS_NAME_2016) %>% 
  group_by(SA3_CODE16, SOSR_CODE_2016) %>% 
  mutate(pop_sa3_16_cat = sum(pop_sa1_16)) %>% 
  ungroup() %>% 
  mutate(pop_sa3_16_s = pop_sa3_16_cat/pop_sa3_16) %>% 
  arrange(SA3_CODE16, desc(pop_sa3_16_s)) %>% 
  group_by(SA3_CODE16, SOSR_CODE_2016) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  select(-SA1_MAIN16, -SA1_7DIG16, -pop_sa1_16) %>% 
  select(SA3_CODE16, SA3_NAME16, everything())

urban_sosr_one <- 
  urban_sosr %>% 
  group_by(SA3_CODE16) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  filter(! SA3_CODE16 %in% c(90101, 90102, 90104, 10803))
```

Share of majority category

```{r}
urban_sosr_one %>% 
  descr(pop_sa3_16_s)

ggplot(urban_sosr_one, aes(pop_sa3_16_s)) + 
  geom_histogram() +
  ylab("Number of SA3") + xlab("Share of majority category") + theme_minimal()
```

Map 

```{r}
SA3 %>% 
  left_join(urban_sosr_one) %>% 
  tm_shape() + 
  tm_polygons(col = "SOSR_NAME_2016", alpha = .5, border.col = NA, id = "SA3_NAME16")
```

### Section of State (SOS)

Original data

```{r}
urban %>% 
  frq(SOS_NAME_2016)
```

SA3 assigned to majority

```{r}
urban_sos <- 
  urban %>% 
  select(-SOSR_CODE_2016, -SOSR_NAME_2016) %>% 
  group_by(SA3_CODE16, SOS_CODE_2016) %>% 
  mutate(pop_sa3_16_cat = sum(pop_sa1_16)) %>% 
  ungroup() %>% 
  mutate(pop_sa3_16_s = pop_sa3_16_cat/pop_sa3_16) %>% 
  arrange(SA3_CODE16, desc(pop_sa3_16_s)) %>% 
  group_by(SA3_CODE16, SOS_CODE_2016) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  select(-SA1_MAIN16, -SA1_7DIG16, -pop_sa1_16) %>% 
  select(SA3_CODE16, SA3_NAME16, everything())

urban_sos_one <- 
  urban_sos %>% 
  group_by(SA3_CODE16) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  filter(! SA3_CODE16 %in% c(90101, 90102, 90104, 10803))
```

Share of majority category

```{r}
urban_sos_one %>% 
  descr(pop_sa3_16_s)

ggplot(urban_sos_one, aes(pop_sa3_16_s)) + 
  geom_histogram() +
  ylab("Number of SA3") + xlab("Share of majority category") + theme_minimal()
```

Map 

```{r}
SA3 %>% 
  left_join(urban_sos_one) %>% 
  tm_shape() + 
  tm_polygons(col = "SOS_NAME_2016", alpha = .5, border.col = NA, id = "SA3_NAME16")
```
