==> R/07_TMR_bt_prep.Rmd <==
---
title: "TMR data preparation"
subtitle: `Travel Times for key priority routes` dataset
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    # css: custom.css
    toc: true
    toc_depth: 4
    toc_float: true
    number_sections: true
    theme: united 
    highlight: pygments 
editor_options: 
  chunk_output_type: console
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.width=8, fig.height=6, dpi=300, out.width="800px", out.height="600px")

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

options(scipen=999)
set.seed(12345)

library(pacman) 
p_load(tidyverse, readr, janitor, visdat, lubridate, sf, tmap, tmaptools)
tmap_mode("view")
```

# Data 

Data comes from TMR via (Open Data Portal)[https://data.qld.gov.au/dataset/travel-times-for-key-priority-routes].

Some notes from TMR:

1. Coordinate system is WGS84
2. 0 travel times means no data collected at that time.
3. Time is AEST

# Locations

```{r message=FALSE}
locations <- read_csv("data/TMR/Average Travel Times/raw/Priority-Route-Link-Details-2019.csv") %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols"))

origins <- locations %>% 
  select(-dest_desc, -dest_longitude, -dest_latitude) %>% 
  st_as_sf(coords = c("origin_longitude", "origin_latitude"), crs = 4326, agr = "constant") %>% 
  rename(desc = origin_desc) %>% 
  mutate(type = "0")

qtm(origins)

destinations <- locations %>% 
  select(-origin_desc, -origin_longitude, -origin_latitude) %>% 
  st_as_sf(coords = c("dest_longitude", "dest_latitude"), crs = 4326, agr = "constant") %>% 
  rename(desc = dest_desc) %>% 
  mutate(type = "1")

lines <- rbind(origins, destinations) %>% 
  arrange(link_details, type) %>% 
  group_by(link_details) %>% 
  dplyr::summarise(from = first(desc), to = last(desc), do_union=FALSE) %>%
  st_cast("LINESTRING") 

tm_shape(lines) +
    tm_lines() + 
  tm_shape(origins) +
    tm_dots(col = "blue", size = 0.08) +
  tm_shape(destinations) +
    tm_dots(col = "red", size = 0.02, shape = 1) 

```

# Time trends 2018

```{r}
times_jan <- read_csv("data/TMR/Average Travel Times/raw/Priority-Route-Bluetooth-Travel-Times-Jan-2018.csv") %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  na_if(0) %>% 
  mutate(interval_end = dmy_hm(interval_end, tz = "Australia/Queensland"))

```

## Missing data patterns

```{r}
times_jan %>% 
  select(starts_with("x")) %>% 
  vis_miss()
```

## Time series

```{r}
times_jan <- times_jan %>% 
  gather(starts_with("x"), key = "link_details", value = "speed")
  
ggplot(times_jan, 
       aes(x=interval_end, y=speed, group=link_details)) +
  geom_line(alpha=0.33) +
  theme_minimal() + xlab("") + ylab("Average speed")
```
