==> R/09_BITRE_prep.Rmd <==
---
title: "BITRE airports data preparation"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    # css: custom.css
    toc: true
    toc_depth: 4
    toc_float: true
    number_sections: true
    theme: united 
    highlight: pygments 
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, 
                      fig.width=9, fig.height=7, dpi=300, out.width="900px", out.height="700px")

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

options(scipen=999)
set.seed(12345)

library(pacman) 
p_load(tidyverse, readr, readxl, magrittr, 
       naniar, janitor, scales, lubridate,  
       skimr, sjmisc, kableExtra, webshot, 
       sf, tmap, tmaptools, ggmap)

tmap_mode("view") # makes map interactive
```

# Data 

## Source 

Data comes from [BITRE](https://www.bitre.gov.au/publications/ongoing/airport_traffic_data.aspx)

```{r}
webshot("https://www.bitre.gov.au/publications/ongoing/airport_traffic_data.aspx", 
        "./data/BITRE/docs/bitre-zoomed.png", zoom = 0.5) # %>% shrink()
```

The release used is `WebMonthlyAirportMay2019.xls`.

```{r message=FALSE}
airport <- read_excel("data/BITRE/raw/WebMonthlyAirportMay2019.xls", 
                      sheet = "Airport Passengers", skip = 6) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  filter(airport != "All Australian Airports") %>% 
  select(-inbound_10, -outbound_11, -starts_with("total_")) %>% 
  mutate(airport = paste0(str_to_title(airport), " Airport")) %>% 
  rename(dom_in = inbound_4, dom_out = outbound_5,
         int_in = inbound_7, int_out = outbound_8) %>%
  mutate(date = ymd(paste0(year, "-", month, "-1"), tz = "Australia/Brisbane")) %>% 
  select(airport, date, everything()) %>% 
  mutate(date = ymd(date))

saveRDS(airport, "data/BITRE/clean/airport.Rds")

airport_2018 <- airport %>% 
  filter(year == 2018) %>%
  group_by(airport) %>% 
  mutate(dom_in_t = sum(dom_in),
         dom_out_t = sum(dom_out),
         int_in_t = sum(int_in),
         int_out_t = sum(int_out)) %>% 
  ungroup() %>% 
  mutate(dom_in_p = (dom_in/dom_in_t)*100,
         dom_out_p = (dom_out/dom_out_t)*100,
         int_in_p = (int_in/int_in_t)*100,
         int_out_p = (int_out/int_out_t)*100) %>%   
  select(airport, date, 
         dom_in, dom_in_t, dom_in_p,
         dom_out, dom_out_t, dom_out_p,
         int_in, int_in_t, int_in_p,
         int_out, int_out_t, int_out_p)

saveRDS(airport_2018, "data/BITRE/clean/airport_2018.Rds")
```

## Additional info 

### Newcastle 

Newcastle is pretty much domestic traffic only. There is some traffic at the end of 2018 though. Accoridng to [wiki](https://en.wikipedia.org/wiki/Newcastle_Airport_(New_South_Wales)#Overview):

*In July 2018 a route between Newcastle and Auckland was announced. It was the only international route operating through Newcastle since 2001. The route, operated by Virgin Australia will be for a limited time, during peak holiday periods. The service will operate three times per week between 22 November 2018 and 17 February 2019. It is expected that if the route is successful, it will continue on a more permanent basis.[6] The airport is leased from the Federal Government for civilian air travel until 2045.[7]*

### Sunshine Coast Airport

Similarly, Sunshine coast has some [temporary international traffic](https://en.wikipedia.org/wiki/Sunshine_Coast_Airport):

*In addition, Air New Zealand serves the city seasonally, with three weekly return flights from Maroochydore to Auckland.*


# Airport locations

## Point

There are `r length(unique(airport$airport))` airports listed in the data. They were geocoded using Google Maps. 

```{r eval=FALSE, include=FALSE}
airport_geo <- tibble::enframe(unique(airport$airport), name = NULL, value = "airport") %>% 
  add_column(lat = NA) %>% 
  add_column(lon = NA) %>% 
  add_column(loctype = NA)

key <- readr::read_file("geocode.txt")

register_google(key = key) 

# geocode("Adelaide Airport, Australia", source = "google", output = "all")
# geocode("Launceston Airport", source = "google", output = "all")
# geocode("Perth Airport", source = "google", output = "all")

for(i in 1:nrow(airport_geo)) {
  
  temp <- geocode(paste0(as.character(airport_geo$airport[i]), ", Australia"), 
                  source = "google", output = "more")
  
  airport_geo$lat[i] <- temp$lat
  airport_geo$lon[i] <- temp$lon
  airport_geo$loctype[i] <- temp$loctype
  
}

airport_geo %<>% 
  st_as_sf(coords = c("lon", "lat"), crs = 4326, remove = FALSE)

# table(st_is_valid(airport_geo))

saveRDS(airport_geo, "data/BITRE/clean/airport_geo.Rds")
```


```{r}
airport_geo <- readRDS("data/BITRE/clean/airport_geo.Rds") %>% 
  st_transform(3112)

airport_geo %>% 
  tm_shape() + 
  tm_dots(text = "airport")
# tm_markers(text = "airport", clustering = FALSE)
```

## SA3 overlay

Simple SA3 overlay might be helpful in case of small airports and remote areas?

```{r}
SA3 <- readRDS("Z:/AIRBNB-Q0931/TEMPO_Airbnb/data/geo/SA3_2016_AUST_clean.rds") %>% 
  select(-AREASQKM16)

SA3_over <- right_join(SA3, 
                       st_join(airport_geo, SA3, join = st_intersects) %>% 
                         select(SA3_CODE16) %>% 
                         st_drop_geometry())

SA3_rest <- anti_join(SA3, SA3_over %>% select(SA3_CODE16) %>% st_drop_geometry())

tm_shape(SA3_over) + 
  tm_polygons(col = "red", alpha = .5, border.col = NA, id = "SA3_NAME16") +
  tm_shape(SA3_rest) + 
  tm_polygons(col = "grey", alpha = .5, border.col = NA, id = "SA3_NAME16") +
  tm_shape(filter(airport_geo, airport == "Adelaide Airport")) + 
  tm_dots(text = "airport")
```

## Areas of 'influence'

```{r include=FALSE}
airport_area <- function(airport_name, airport_SA3, airport_influence_type, airport_influence_value) {
  
    # SA3 overlaping airport is red
    tm_shape(filter(SA3_over, SA3_NAME16 == airport_SA3)) +
    tm_polygons(col = "red", alpha = .5, border.col = NA, id = "SA3_NAME16") +
    
    # other SA3s with airports are darkgrey
    tm_shape(filter(SA3_over, SA3_NAME16 != airport_SA3)) +
    tm_polygons(col = "darkgrey", alpha = .5, border.col = NA, id = "SA3_NAME16") +
    
    # main airport is larger
    tm_shape(filter(airport_geo, airport == airport_name)) + 
    tm_dots(size = 0.05, text = "airport") +
    
    # other airports are smaller
    tm_shape(filter(airport_geo, airport != airport_name)) + 
    tm_dots(size = 0.01, text = "airport") +
    
    # zone of influence is in orange
    # areas with no airports and not in zone of influence
    if (airport_influence_type == "SA4_NAME16")  {
      tm_shape(filter(SA3_rest, SA4_NAME16 == airport_influence_value)) +
        tm_polygons(col = "orange", alpha = .5, border.col = NA, id = "SA3_NAME16") +
      tm_shape(filter(SA3_rest, SA4_NAME16 != airport_influence_value)) +
        tm_polygons(col = "lightgrey", alpha = .5, border.col = NA, id = "SA3_NAME16")
    } 
    else if (airport_influence_type == "GCC_NAME16") {
      tm_shape(filter(SA3_rest, GCC_NAME16 == airport_influence_value)) +
        tm_polygons(col = "orange", alpha = .5, border.col = NA, id = "SA3_NAME16")  +
      tm_shape(filter(SA3_rest, GCC_NAME16 != airport_influence_value)) +
        tm_polygons(col = "lightgrey", alpha = .5, border.col = NA, id = "SA3_NAME16")
    } 
    else if (airport_influence_type == "none") {
      tm_shape(SA3_rest) +
        tm_polygons(col = "lightgrey", alpha = .5, border.col = NA, id = "SA3_NAME16")
    } 
  }
```

```{r}
# sjmisc::frq(airport_geo$airport)

airport_area(airport_name = "Adelaide Airport", airport_SA3 = "West Torrens",
             airport_influence_type = "GCC_NAME16", airport_influence_value = "Greater Adelaide")

airport_area(airport_name = "Alice Springs Airport", airport_SA3 = "Alice Springs",
             airport_influence_type = "none")

airport_area(airport_name = "Ballina Airport", airport_SA3 = "Richmond Valley - Coastal",
             airport_influence_type = "none")

airport_area(airport_name = "Brisbane Airport", airport_SA3 = "Nundah",
             airport_influence_type = "GCC_NAME16", airport_influence_value = "Greater Brisbane")

airport_area(airport_name = "Cairns Airport", airport_SA3 = "Cairns - North",
             airport_influence_type = "SA4_NAME16", airport_influence_value = "Cairns")

airport_area(airport_name = "Canberra Airport", airport_SA3 = "Canberra East",
             airport_influence_type = "GCC_NAME16", airport_influence_value = "Australian Capital Territory")

airport_area(airport_name = "Darwin Airport", airport_SA3 = "Darwin City",
             airport_influence_type = "GCC_NAME16", airport_influence_value = "Greater Darwin")

airport_area(airport_name = "Gold Coast Airport", airport_SA3 = "Coolangatta",
             airport_influence_type = "SA4_NAME16", airport_influence_value = "Gold Coast")

airport_area(airport_name = "Hobart Airport", airport_SA3 = "Hobart - North East",
             airport_influence_type = "GCC_NAME16", airport_influence_value = "Greater Hobart")

airport_area(airport_name = "Karratha Airport", airport_SA3 = "West Pilbara",
             airport_influence_type = "none")

airport_area(airport_name = "Launceston Airport", airport_SA3 = "North East",
             airport_influence_type = "SA4_NAME16", airport_influence_value = "Launceston and North East")

airport_area(airport_name = "Mackay Airport", airport_SA3 = "Mackay",
             airport_influence_type = "SA4_NAME16", airport_influence_value = "Mackay - Isaac - Whitsunday")

airport_area(airport_name = "Melbourne Airport", airport_SA3 = "Tullamarine - Broadmeadows",
             airport_influence_type = "GCC_NAME16", airport_influence_value = "Greater Melbourne")

airport_area(airport_name = "Newcastle Airport", airport_SA3 = "Port Stephens",
             airport_influence_type = "SA4_NAME16", airport_influence_value = "Newcastle and Lake Macquarie")

airport_area(airport_name = "Perth Airport", airport_SA3 = "Belmont - Victoria Park",
             airport_influence_type = "GCC_NAME16", airport_influence_value = "Greater Perth")

airport_area(airport_name = "Proserpine Airport", airport_SA3 = "Whitsunday",
             airport_influence_type = "none")

airport_area(airport_name = "Rockhampton Airport", airport_SA3 = "Rockhampton",
             airport_influence_type = "none")

airport_area(airport_name = "Sunshine Coast Airport", airport_SA3 = "Maroochy",
             airport_influence_type = "SA4_NAME16", airport_influence_value = "Sunshine Coast")

airport_area(airport_name = "Sydney Airport", airport_SA3 = "Botany",
             airport_influence_type = "GCC_NAME16", airport_influence_value = "Greater Sydney")

airport_area(airport_name = "Townsville Airport", airport_SA3 = "Townsville",
             airport_influence_type = "SA4_NAME16", airport_influence_value = "Townsville")

```

## Closest international airport

Note Newcastle, Sunshine Coast & Townsville are treated as domestic for this calculations.

```{r}

```



# Time trends 2009+

## Domestic inbound 

```{r}
airport %>% 
  ggplot(aes(x = date, y = dom_in, group = airport, colour = airport)) +
  geom_line() + 
  theme_light() + 
  ylab("Domestic inbound passengers") + xlab("")
```

## International outbound

```{r}
airport %>% 
  filter(int_out > 0) %>% 
  ggplot(aes(x = date, y = int_out, group = airport, colour = airport)) +
  geom_line() + 
  theme_light() + 
  ylab("Int'l outbound passengers") + xlab("")
```

# 2018 data

## Absolute numbers

### Int'l 

```{r}
airport_2018 %>% 
  filter(airport %in% c("Brisbane Airport", "Melbourne Airport", "Sydney Airport", 
                        "Perth Airport", "Adelaide Airport", "Gold Coast Airport",
                        "Cairns Airport", "Canberra Airport", "Darwin Airport")) %>%
  gather(dom_in:int_out_p, key = "type", value = "passengers") %>% 
  separate(type, sep = "_", c("source", "bound", "measure")) %>% 
  filter(is.na(measure)) %>% 
  ggplot(aes(x = date, y = passengers, fill = interaction(source, bound))) +
  geom_col() + 
  facet_wrap(~airport, scales = "fixed") +
  theme_light() + 
  ylab("No. of passengers") + xlab("")
```

### Regional

```{r}
airport_2018 %>% 
  filter(! airport %in% c("Brisbane Airport", "Melbourne Airport", "Sydney Airport", 
                        "Perth Airport", "Adelaide Airport", "Gold Coast Airport",
                        "Cairns Airport", "Canberra Airport", "Darwin Airport")) %>%
  gather(dom_in:int_out_p, key = "type", value = "passengers") %>% 
  separate(type, sep = "_", c("source", "bound", "measure")) %>% 
  filter(is.na(measure)) %>% 
  ggplot(aes(x = date, y = passengers, fill = interaction(source, bound))) +
  geom_col() + 
  facet_wrap(~airport, scales = "fixed") +
  theme_light() + 
  ylab("No. of passengers") + xlab("")
```

## Relative numbers

```{r}
airport_2018 %>% 
  # filter(airport %in% c("Brisbane Airport", "Melbourne Airport", "Perth Airport", "Sydney Airport", "Adelaide Airport", "Gold Coast Airport")) %>%
  gather(dom_in:int_out_p, key = "type", value = "passengers") %>% 
  separate(type, sep = "_", c("source", "bound", "measure")) %>% 
  filter(measure == "p") %>% 
  ggplot(aes(x = date, y = passengers, colour = source, linetype = bound)) +
  geom_line() + 
  facet_wrap(~airport, scales = "fixed") +
  theme_light() + 
  ylab("% of 2018 passengers") + xlab("")

```

Clearer trends if Newcastle, Sunshine Coast & Townsville are treated as domestic only

```{r}
airport_2018 %>% 
  gather(dom_in:int_out_p, key = "type", value = "passengers") %>% 
  separate(type, sep = "_", c("source", "bound", "measure")) %>% 
  filter(measure == "p") %>% 
  filter(!(airport == "Newcastle Airport" & source == "int")) %>%
  filter(!(airport == "Sunshine Coast Airport" & source == "int")) %>%
  filter(!(airport == "Townsville Airport" & source == "int")) %>%
  ggplot(aes(x = date, y = passengers, colour = source, linetype = bound)) +
  geom_line() + 
  facet_wrap(~airport, scales = "fixed") +
  theme_light() + 
  ylab("% of 2018 passengers") + xlab("")

```
