==> R/14_data_comb.Rmd <==
---
title: "Combining data sources"
# subtitle: 
# author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    highlight: pygments
    keep_md: no
    # number_sections: yes
    theme: united
    # toc: yes
    toc_depth: 4
    toc_float: yes
editor_options: 
  chunk_output_type: console
---

<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, 
                      fig.width=9, fig.height=7, out.width="900px", out.height="700px", dpi=300)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

options(scipen=999)
set.seed(12345)

library(pacman) 
p_load(tidyverse, readr, lubridate)
```

<!-- ------------------------------------------------------------ --> 

# Pop present

```{r include=FALSE}
pop_present_au_18_month_long <- 
  readRDS("Data/ABS/pop_present/pop_present_month.Rds") %>% 
  select(-end) %>% 
  rename(date = start)

write_csv(pop_present_au_18_month_long, "data/model_input/pop_present_au_18_month_long.csv")

# pop_present_au_18_wide <- 
#   pop_present_au_18_month_long %>% 
#   mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
#   # mutate(SA3_CODE16 = NA, SA3_NAME16 = "") %>% 
#   dplyr::select(-date) %>% 
#   pivot_wider(names_from = month, values_from = mean_month_pop_present)
# 
# write_csv(pop_present_au_18_wide, "data/model_input/pop_present_au_18_wide.csv")
```

```{r}
summary(pop_present_au_18_month_long)
```

# ERP

## Single estimate

```{r include=FALSE}
erp_sa3_18_year_long <- readRDS("./data/ABS/3218.0 Regional Population Growth, Australia, 2017-18/clean/ERP_SA2_2018.Rds") %>% 
  group_by(sa3_code, sa3_name) %>% 
  summarise(erp = sum(ERP_18)) %>% 
  ungroup() %>% 
  rename(SA3_CODE16 = sa3_code, SA3_NAME16 = sa3_name) %>% 
  mutate(date = ymd("2018-06-30")) %>% 
  select(date, everything())

write_csv(erp_sa3_18_year_long, "data/model_input/erp_sa3_18_year_long.csv")

# erp_sa3_18_wide <- 
#   erp_sa3_18_year_long %>% 
#   mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
#   select(-date) %>% 
#   pivot_wider(names_from = month, values_from = erp)
# 
# write_csv(erp_sa3_18_wide, "data/model_input/erp_sa3_18_wide.csv")
```

```{r}
summary(erp_sa3_18_year_long)
```

## Interpolated to months

```{r}
erp_sa3_18_month_long_adj <- read_csv("data/model_input/erp_sa3_18_month_adj_long.csv")
```

```{r}
summary(erp_sa3_18_month_long_adj)
```

<!-- ------------------------------------------------------------ --> 

# IVS

```{r}
IVS_SA3_18qtr_long <- read_csv("data/model_input/IVS_SA3_18qtr_long.csv") %>% 
  mutate(qtr = date)

IVS_sa3_18_month_long <- 
  expand(IVS_SA3_18qtr_long, nesting(SA3_CODE16, SA3_NAME16), 
         date = seq(from = ymd("2018-01-01"), to = ymd("2018-12-01"), by = "month")) %>% 
  left_join(IVS_SA3_18qtr_long) %>% 
  fill(IVS_visitor_nights, IVS_sample_count, qtr, .direction = "up") %>% 
  mutate(days_in_month = days_in_month(date)) %>% 
  group_by(SA3_CODE16, qtr) %>% 
  mutate(days_in_qtr = sum(days_in_month)) %>% 
  ungroup() %>% 
  mutate(share = days_in_month / days_in_qtr) %>% 
  mutate(IVS_visitor_nights = IVS_visitor_nights * share,
         IVS_sample_count = IVS_sample_count * share) %>% 
  select(-share, -qtr, -days_in_month, -days_in_qtr)

rm(IVS_SA3_18qtr_long)

write_csv(IVS_sa3_18_month_long, "data/model_input/IVS_sa3_18_month_long.csv")
```

Quarters interpolated to months using share of days of each months in a quarter.

```{r}
summary(IVS_sa3_18_month_long)
```

<!-- ------------------------------------------------------------ --> 

# NVS

## Origin

```{r}
NVS_origin_SA3_18_month_long <- read_csv("data/model_input/NVS_origin_SA3_18_month_long.csv")
```

## Destination

```{r}
NVS_dest_SA3_18_month_long <- read_csv("data/model_input/NVS_dest_SA3_18_month_long.csv")
```

## Int'l outbound

```{r}
NVS_intl_outbound_SA3_18qtr_Dec17_long <- read_csv("data/model_input/NVS_intl_outbound_SA3_18qtr+Dec17_long.csv") %>% 
  mutate(qtr = date)

NVS_intl_outbound_sa3_18_month_long <- 
  expand(NVS_intl_outbound_SA3_18qtr_Dec17_long, nesting(SA3_CODE16, SA3_NAME16), 
         date = seq(from = ymd("2018-01-01"), to = ymd("2018-12-01"), by = "month")) %>% 
  left_join(NVS_intl_outbound_SA3_18qtr_Dec17_long) %>% 
  fill(NVS_intl_out_visitor_nights, NVS_intl_out_sample_count, qtr, .direction = "up") %>% 
  mutate(days_in_month = days_in_month(date)) %>% 
  group_by(SA3_CODE16, qtr) %>% 
  mutate(days_in_qtr = sum(days_in_month)) %>% 
  ungroup() %>% 
  mutate(share = days_in_month / days_in_qtr) %>% 
  mutate(NVS_intl_out_visitor_nights = NVS_intl_out_visitor_nights * share,
         NVS_intl_out_sample_count = NVS_intl_out_sample_count * share) %>% 
  select(-share, -qtr, -days_in_month, -days_in_qtr)

rm(NVS_intl_outbound_SA3_18qtr_Dec17_long)

write_csv(NVS_intl_outbound_sa3_18_month_long, "data/model_input/NVS_intl_outbound_sa3_18_month_long.csv")
```

Quarters interpolated to months using share of days of each months in a quarter.

```{r}
summary(NVS_intl_outbound_sa3_18_month_long)
```

<!-- ------------------------------------------------------------ --> 

# BITRE

```{r}
bitre_airport_18_month_long <- readRDS("data/BITRE/clean/airport_2018.Rds")

write_csv(bitre_airport_18_month_long, "data/model_input/bitre_airport_18_month_long.csv")
```

```{r}
summary(bitre_airport_18_month_long)
```

<!-- ------------------------------------------------------------ --> 

# Airbnb

## Utilization

Using (maximum) number of guest nights occupied (GNO).

```{r include=FALSE}
airbnb_sa3_18_month_long <- 
  readRDS(file = "Z:/AIRBNB-Q0931/TEMPO_Airbnb/data/airdna/clean/airbnb_sa3_18.rds") %>% 
  dplyr::select(SA3_CODE16, SA3_NAME16, reporting_month, max_gno, max_gna) %>% 
  dplyr::rename(date = reporting_month, airbnb_gno = max_gno, airbnb_gna = max_gna)

write_csv(airbnb_sa3_18_month_long, "data/model_input/airbnb_sa3_18_month_long.csv")

# airbnb_sa3_18_wide <- 
#   airbnb_sa3_18_month_long %>% 
#   mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
#   select(-date) %>% 
#   pivot_wider(names_from = month, values_from = airbnb_gno)
# 
# write_csv(airbnb_sa3_18_wide, "data/model_input/airbnb_sa3_18_wide.csv")
```

```{r}
summary(airbnb_sa3_18_month_long)
```

## Supply

Using (maximum) number of guest nights occupied (GNO).

```{r include=FALSE}
airbnb_sa3_18_month_long_supply <- 
  readRDS(file = "Z:/AIRBNB-Q0931/TEMPO_Airbnb/data/airdna/clean/airbnb_sa3_18.rds") %>% 
  dplyr::select(SA3_CODE16, SA3_NAME16, reporting_month, starts_with("supply")) 

write_csv(airbnb_sa3_18_month_long_supply, "data/model_input/airbnb_sa3_18_month_long_supply.csv")

# airbnb_sa3_18_wide <- 
#   airbnb_sa3_18_month_long %>% 
#   mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
#   select(-date) %>% 
#   pivot_wider(names_from = month, values_from = airbnb_gno)
# 
# write_csv(airbnb_sa3_18_wide, "data/model_input/airbnb_sa3_18_wide.csv")
```

```{r}
summary(airbnb_sa3_18_month_long_supply)
```


<!-- ------------------------------------------------------------ --> 

# HomeAway

## Utilization

Using (maximum) number of guest nights occupied (GNO).

```{r include=FALSE}
ha_sa3_18_month_long <- 
  readRDS(file = "Z:/AIRBNB-Q0931/TEMPO_Airbnb/data/ha/clean/ha_sa3_18.rds") %>% 
  dplyr::select(SA3_CODE16, SA3_NAME16, reporting_month, max_gno, max_gna) %>% 
  dplyr::rename(date = reporting_month, ha_gno = max_gno, ha_gna = max_gna)

write_csv(ha_sa3_18_month_long, "data/model_input/ha_sa3_18_month_long.csv")

# ha_sa3_18_wide <- ha_sa3_18_month_long %>% 
#   mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
#   select(-date) %>% 
#   pivot_wider(names_from = month, values_from = ha_gno)
# 
# write_csv(ha_sa3_18_wide, "data/model_input/ha_sa3_18_wide.csv")
```

```{r}
summary(ha_sa3_18_month_long)
```

## Supply

Using (maximum) number of guest nights occupied (GNO).

```{r include=FALSE}
ha_sa3_18_month_long_supply <- 
  readRDS(file = "Z:/AIRBNB-Q0931/TEMPO_Airbnb/data/ha/clean/ha_sa3_18.rds") %>% 
  dplyr::select(SA3_CODE16, SA3_NAME16, reporting_month, starts_with("supply")) 

write_csv(ha_sa3_18_month_long_supply, "data/model_input/ha_sa3_18_month_long_supply.csv")

# ha_sa3_18_month_wide_supply <- 
#   ha_sa3_18_month_long_supply %>% 
#   mutate(month = month(date, label = TRUE, abbr = TRUE)) %>% 
#   select(-date) %>% 
#   pivot_wider(names_from = month, values_from = airbnb_gno)
# 
# write_csv(ha_sa3_18_month_wide_supply, "data/model_input/ha_sa3_18_month_wide_supply.csv")
```

```{r}
summary(ha_sa3_18_month_long_supply)
```

<!-- ------------------------------------------------------------ --> 

# Master file

```{r}
tempo_sa3_18_month_long <- 
  erp_sa3_18_month_long_adj %>% 
  left_join(NVS_intl_outbound_sa3_18_month_long) %>% 
  left_join(NVS_origin_SA3_18_month_long) %>% 
  left_join(NVS_dest_SA3_18_month_long) %>% 
  left_join(IVS_sa3_18_month_long) %>% 
  left_join(airbnb_sa3_18_month_long) %>% 
  left_join(ha_sa3_18_month_long) %>% 
  select(-contains("sample")) %>% 
  rename(Airbnb_visitor_nights = airbnb_gno,
         Homeaway_visitor_nights = ha_gno) %>% 
  arrange(SA3_CODE16, date) %>% 
  filter(! SA3_NAME16 %in% c("Lord Howe Island")) 

write_csv(tempo_sa3_18_month_long, "data/model_input/tempo_sa3_18_month_long.csv")
saveRDS(tempo_sa3_18_month_long, "data/model_input/tempo_sa3_18_month_long.Rds")

summary(tempo_sa3_18_month_long)
```
