==> R/18_ml_ext.Rmd <==
---
title: "Airbnb & NVS models"
subtitle: "Experiments with alternative GLMs"
# author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    highlight: pygments
    keep_md: no
    # number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
editor_options: 
  chunk_output_type: console
---

<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      message = FALSE, warning = FALSE, 
                      fig.width = 8, fig.height = 5,
                      out.width = "800px", out.height = "500px",
                      dpi = 300)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

options(scipen=999)
set.seed(12345)

library(pacman) 
p_load(tidyverse, lubridate)

```

```{r data, include=FALSE}
data <- readRDS("data/model_input/tempo_sa3_18_month_long.Rds")

glimpse(data)

# length(unique(data$SA3_CODE16))
# length(unique(data$SA3_NAME16))

# `Illawarra Catchment Reserve` has no Airbnb data. TS filled with zeros.

data <- data %>% 
  mutate(Airbnb_visitor_nights = ifelse(is.na(Airbnb_visitor_nights), 
                                        0, Airbnb_visitor_nights))

aggregated <- data %>% 
  group_by(SA3_CODE16, SA3_NAME16) %>% 
  summarise(NVS_dest_nights_year = as.integer(round(sum(NVS_dest_nights))), 
            Airbnb_visitor_nights_year = as.integer(round(sum(Airbnb_visitor_nights)))) %>%   
  ungroup()

data <- data %>% 
  left_join(aggregated)

data <- data %>% 
  group_by(SA3_CODE16) %>% 
  mutate(time = row_number()) %>% 
  ungroup()

regions <- readRDS("./data/regions/regions.rds") %>% 
  dplyr::select(sa3, region_types)  %>% 
  rename(SA3_NAME16 = sa3)

climate <- read_csv("data/model_input/climate_zones_SA3_centroid.csv", 
                    col_types = cols(SA3_NAME_2016 = col_skip(), 
                                     cz_code = col_integer())) %>% 
  rename(SA3_CODE16 = SA3_CODE_2016)

data <- data %>% 
  select(-erp_adj, -NVS_intl_out_visitor_nights, -NVS_origin_nights,
         -airbnb_gna, -Homeaway_visitor_nights, -ha_gna) %>% 
  left_join(climate)

remote <- read_csv("data/model_input/remote_sa3_18.csv") %>% 
  select(SA3_CODE16, RA_NAME_2016) %>% 
  mutate(RA_NAME_2016 = ifelse(RA_NAME_2016 == "Very Remote Australia", 
                               "Remote Australia", RA_NAME_2016))

data <- data %>% 
  left_join(remote)

data$SA3_NAME16 <- as_factor(data$SA3_NAME16)
data$cz_desc <- as_factor(data$cz_desc)
data$RA_NAME_2016 <- as_factor(data$RA_NAME_2016)

rm(remote, climate)
```

# Testing `mgcv` v1

```{r}
p_load(mgcv, gratia)
```

## Cyclical time only

```{r}
m0 <- gam(Airbnb_visitor_nights ~ 1 + 
            s(time, bs = "cc"),
          offset = log(NVS_dest_nights_year),
          data = data, 
          # niterPQL=500,
          # family = negbin(1))
          family = poisson)

summary(m0) 

# gam.check(m0)
# plot.gam(m0, shade = TRUE)

draw(m0)
appraise(m0)

# acf(residuals(m0), main = "Raw residual ACF")
# acf(residuals(m0, type = "normalized"), main = "Standardized residual ACF")

data$m0 <- fitted(m0)
# data$m0 <- predict.gam(m0, type = "response")

head(predict.gam(m0, type="link"))

str(predict.gam(m0, type="terms"))
dim(predict.gam(m0, type="terms"))
head(predict.gam(m0, type="terms"), n =24)

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m0),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m0),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

```

```{r}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m0, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

```{r eval=FALSE, include=FALSE}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m0, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  geom_point(aes(y = NVS_dest_nights),
             col = "#E7B800", alpha = 0.33) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

## Cyclical time by climate region

```{r}
# https://www.fromthebottomoftheheap.net/2017/12/14/difference-splines-ii/
# time needed if interaction included?

m1 <- gam(Airbnb_visitor_nights ~ 1 + 
            cz_desc +
            s(time, bs = "cc") +
            s(time, by = cz_desc, bs = "cc"),
          offset = log(NVS_dest_nights_year),
          data = data, 
          # niterPQL=500,
          # family = negbin(1))
          family = poisson)

summary(m1) 

# gam.check(m1)
# plot.gam(m1, shade = TRUE, pages = 1, scale = 0, seWithMean = TRUE)

draw(m1)
appraise(m1)

# acf(residuals(m1), main = "Raw residual ACF")
# acf(residuals(m1, type = "normalized"), main = "Standardized residual ACF")

data$m1 <- fitted(m1)
# data$m1 <- predict.gam(m1, type = "response")

head(predict.gam(m1, type="link"))

str(predict.gam(m1, type="terms"))
dim(predict.gam(m1, type="terms"))
head(predict.gam(m1, type="terms"), n =24)

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m1),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m1),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m1),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~cz_desc) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

```

```{r}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m1, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

```{r eval=FALSE, include=FALSE}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m1, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  geom_point(aes(y = NVS_dest_nights),
             col = "#E7B800", alpha = 0.33) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

## Cyclical time by climate region + random effects on SA3

```{r}
m2 <- gam(Airbnb_visitor_nights ~ 1 + 
            cz_desc +
            s(time, bs = "cc") +
            s(time, by = cz_desc, bs = "cc") +
            s(SA3_CODE16, bs="re", k = length(unique(data$SA3_NAME16))),
          offset = log(NVS_dest_nights_year),
          data = data, 
          # niterPQL=500,
          # family = negbin(1))
          family = poisson)

summary(m2) 

# gam.check(m2)
# plot.gam(m2, shade = TRUE, pages = 1, scale = 0, seWithMean = TRUE)

draw(m2)
appraise(m2)

# acf(residuals(m2), main = "Raw residual ACF")
# acf(residuals(m2, type = "normalized"), main = "Standardized residual ACF")

data$m2 <- fitted(m2)
# data$m2 <- predict.gam(m2, type = "response")

head(predict.gam(m2, type="link"))

str(predict.gam(m2, type="terms"))
dim(predict.gam(m2, type="terms"))
head(predict.gam(m2, type="terms"), n =24)

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m2),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m2),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m2),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~cz_desc) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

```

```{r}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m2, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

```{r eval=FALSE, include=FALSE}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m2, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  geom_point(aes(y = NVS_dest_nights),
             col = "#E7B800", alpha = 0.33) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```


## Cyclical time by climate region + cyclical time by RA, random effects on SA3

```{r}
m3 <- gam(Airbnb_visitor_nights ~ 1 + 
            s(time, bs = "cc") +
            cz_desc +
            s(time, by = cz_desc, bs = "cc") +
            RA_NAME_2016 +
            s(time, by = RA_NAME_2016, bs = "cc") +
            s(SA3_CODE16, bs="re", k = length(unique(data$SA3_NAME16))),
          offset = log(NVS_dest_nights_year),
          data = data, 
          # niterPQL=500,
          # family = negbin(1))
          family = poisson)

summary(m3) 

# gam.check(m3)
# plot.gam(m3, shade = TRUE, pages = 1, scale = 0, seWithMean = TRUE)

# draw(m3)
appraise(m3)

# acf(residuals(m3), main = "Raw residual ACF")
# acf(residuals(m3, type = "normalized"), main = "Standardized residual ACF")

data$m3 <- fitted(m3)
# data$m3 <- predict.gam(m3, type = "response")

head(predict.gam(m3, type="link"))

str(predict.gam(m3, type="terms"))
dim(predict.gam(m3, type="terms"))
head(predict.gam(m3, type="terms"), n =24)

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m3),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m3),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m3),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~cz_desc) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m3),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~RA_NAME_2016) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

```

```{r}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m3, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

```{r}
data %>% 
  ggplot(aes(x = time, y = m3, group = SA3_NAME16)) +
  geom_line(aes(group = SA3_NAME16), alpha = 0.3) +
  geom_smooth(aes(group = cz_desc)) + 
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~cz_desc, scales = "free_y")
```

```{r}
data %>% 
  ggplot(aes(x = time, y = m3, group = SA3_NAME16)) +
  geom_line(aes(group = SA3_NAME16), alpha = 0.3) +
  geom_smooth(aes(group = RA_NAME_2016)) + 
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~RA_NAME_2016, scales = "free_y")
```

```{r}
data %>% 
  ggplot(aes(x = time, y = m3, group = SA3_NAME16)) +
  geom_line(aes(group = SA3_NAME16), alpha = 0.3) +
  geom_smooth(aes(group = RA_NAME_2016)) + 
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_grid(cz_desc~RA_NAME_2016, scales = "free_y")
```

```{r eval=FALSE, include=FALSE}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m3, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  geom_point(aes(y = NVS_dest_nights),
             col = "#E7B800", alpha = 0.33) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

## Cyclical time by climate region + RA interaction, random effects on SA3

```{r}
# https://stackoverflow.com/questions/47934100/how-to-specify-the-non-linear-interaction-of-two-factor-variables-in-generalised

m4 <- gam(Airbnb_visitor_nights ~ 1 + 
            cz_desc + RA_NAME_2016 +
            s(time, bs = "cc") +
            s(time, by = interaction(cz_desc, RA_NAME_2016), bs = "cc") +
            s(SA3_CODE16, bs="re", k = length(unique(data$SA3_NAME16))),
          offset = log(NVS_dest_nights_year),
          data = data, 
          # niterPQL=500,
          # family = negbin(1))
          family = poisson)

summary(m4) 

# gam.check(m4)
# plot.gam(m4, shade = TRUE, pages = 1, scale = 0, seWithMean = TRUE)

# draw(m4)
appraise(m4)

# acf(residuals(m4), main = "Raw residual ACF")
# acf(residuals(m4, type = "normalized"), main = "Standardized residual ACF")

data$m4 <- fitted(m4)
# data$m4 <- predict.gam(m4, type = "response")

head(predict.gam(m4, type="link"))

str(predict.gam(m4, type="terms"))
dim(predict.gam(m4, type="terms"))
head(predict.gam(m4, type="terms"), n =24)

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m4),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m4),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m4),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~cz_desc) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m4),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~RA_NAME_2016) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

```

```{r}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m4, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

```{r}
data %>% 
  ggplot(aes(x = time, y = m4, group = SA3_NAME16)) +
  geom_line(aes(group = SA3_NAME16), alpha = 0.3) +
  geom_smooth(aes(group = cz_desc)) + 
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~cz_desc, scales = "free_y")
```

```{r}
data %>% 
  ggplot(aes(x = time, y = m4, group = SA3_NAME16)) +
  geom_line(aes(group = SA3_NAME16), alpha = 0.3) +
  geom_smooth(aes(group = RA_NAME_2016)) + 
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~RA_NAME_2016, scales = "free_y")
```

```{r}
data %>% 
  ggplot(aes(x = time, y = m4, group = SA3_NAME16)) +
  geom_line(aes(group = SA3_NAME16), alpha = 0.3) +
  geom_smooth(aes(group = RA_NAME_2016)) + 
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_grid(cz_desc~RA_NAME_2016, scales = "free_y")
```

```{r eval=FALSE, include=FALSE}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m4, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  geom_point(aes(y = NVS_dest_nights),
             col = "#E7B800", alpha = 0.33) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")

```
