==> R/19_ml_experiment.Rmd <==
---
title: "Airbnb & NVS models"
subtitle: "Experiments with alternative GLMs"
# author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    highlight: pygments
    keep_md: no
    # number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
editor_options: 
  chunk_output_type: console
---

<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      message = FALSE, warning = FALSE, 
                      fig.width = 8, fig.height = 5,
                      out.width = "800px", out.height = "500px",
                      dpi = 300)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

options(scipen=999)
set.seed(12345)

library(pacman) 
p_load(tidyverse, lubridate)

```

# Testing `mgcv` v2

```{r}
p_load(mgcv)
```

## Cyclical time only

```{r}
m1 <- gamm(Airbnb_visitor_nights ~ 1 + 
             # s(time, bs = "tp"),
             # s(time, bs = "cr"),
             s(time, bs = "cc"),
           offset = log(NVS_dest_nights_year),
           random=list(SA3_NAME16 = ~ 1),
           data = data, 
           niterPQL=500,
           # family = negbin(1))
           family = poisson)

summary(m1$lme) # details of underlying lme fit
summary(m1$gam) # gam style summary of fitted model

gam.check(m1$gam)
plot.gam(m1$gam, shade = TRUE)

data$m1 <- exp(fitted(m1$lme))

acf(residuals(m1$gam), main = "Raw residual ACF")
acf(residuals(m1$lme, type = "normalized"), main = "Standardized residual ACF")

predict.gam(m1$gam, type="link")
predict.gam(m1$gam, type = "response")
predict.gam(m1$gam, type="terms")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m1),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m1),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")


```

## Cyclical time by SA3

```{r}
m2 <- gamm(Airbnb_visitor_nights ~ 1 + 
             # s(time, bs = "tp"),
             # s(time, bs = "cr"),
             s(time, bs = "cc", by = SA3_NAME16),
           offset = log(NVS_dest_nights_year),
           random=list(SA3_NAME16 = ~ 1),
           data = data, 
           niterPQL=500,
           family = negbin(1))
# family = poisson)

summary(m2$lme) # details of underlying lme fit
summary(m2$gam) # gam style summary of fitted model

gam.check(m2$gam)
plot.gam(m2$gam, shade = TRUE)

data$m2 <- exp(fitted(m2$lme))

acf(residuals(m2$gam), main = "Raw residual ACF")
acf(residuals(m2$lme, type = "normalized"), main = "Standardized residual ACF")

predict.gam(m2$gam, type="link")
predict.gam(m2$gam, type = "response")
predict.gam(m2$gam, type="terms")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m2),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m2),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")


```
## Example areas

```{r}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m1, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```

```{r}
data %>% 
  right_join(regions) %>% 
  ggplot(aes(x = time)) +
  geom_line(aes(y = m1, group = SA3_NAME16, col = region_types)) +
  geom_point(aes(y = Airbnb_visitor_nights), 
             col = "#00AFBB", alpha = 0.3) +
  geom_point(aes(y = NVS_dest_nights),
             col = "#E7B800", alpha = 0.33) +
  theme_minimal() + 
  scale_x_continuous(limits = c(0, 12), breaks = c(2:10)) +
  xlab("") + ylab("Observed/fitted") + 
  facet_wrap(~SA3_NAME16, scales = "free_y")
```


# Testing `gamm4`

```{r}
p_load(gamm4)
```

## Cyclical time only

```{r}
m1 <- gamm4(Airbnb_visitor_nights ~ 1 + 
              # s(time, bs = "tp"),
              # s(time, bs = "cr"),
              s(time, bs = "cc"),
            random = ~(1 | SA3_NAME16),
            data = data, 
            family = poisson)

summary(m1$mer)
summary(m1$gam)
plot(m1$gam, shade=TRUE)

data$m1 <- predict(m1$gam, type = "response")
data$m1 <- fitted(m1$gam)

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m1),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m1),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")
```

## Cyclical time by climate zone

```{r}
m2 <- gamm4(Airbnb_visitor_nights ~ 1 + 
              # s(time, bs = "tp"),
              # s(time, bs = "cr", by = SA3_NAME16),
              # s(time, bs = "cc", by = SA3_NAME16),
              s(time, bs = "cc", by = cz_desc),
            random = ~(1 | SA3_NAME16),
            data = data, 
            family = poisson)

summary(m2$mer)
summary(m2$gam)
plot(m2$gam, shade=TRUE, rug=TRUE, pages=1)

data$m2 <- predict(m2$gam, type = "response")
data$m1 <- fitted(m2$gam)

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m2),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m2),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")
```



# Testing `glmmADMB`

```{r}
p_load(glmmADMB)

data$SA3_NAME16 <- as_factor(data$SA3_NAME16)

m1 <- glmmadmb(Airbnb_visitor_nights ~ 1 + 
                 offset(log(NVS_dest_nights_year)) +
                 (1 | SA3_NAME16),
               data = data, 
               # family = "poisson")
               family = "nbinom")

m2 <- glmmadmb(Airbnb_visitor_nights ~ 1 + time + 
                 offset(log(NVS_dest_nights_year)) +
                 (1 | SA3_NAME16),
               data = data, 
               # family = "poisson")
               family = "nbinom")

m3 <- glmmadmb(Airbnb_visitor_nights ~ 1 + time + 
                 offset(log(NVS_dest_nights_year)) +
                 (1 + time | SA3_NAME16),
               data = data, 
               # family = "poisson")
               family = "nbinom")





# DHARMa to check residuals
p_load("DHARMa")

# Generate simulated residuals

simulationOutput <- simulateResiduals(fittedModel = m1, n = 1000)

# Plot simulated residuals

plotSimulatedResiduals(simulationOutput = simulationOutput)

par(mfrow = c(1,3))
plotResiduals(scaled.mydata$P_NONWHT,  simulationOutput$scaledResiduals)
plotResiduals(scaled.mydata$a_hinc,  simulationOutput$scaledResiduals)
plotResiduals(scaled.mydata$P_NONWHT*scaled.mydata$a_hinc,  simulationOutput$scaledResiduals)

# Tests

# K-S test for uniformity of scaled residuals; significant = cannot reject non-uniformity (i.e. evidence of non-uniformity)

testUniformity(simulationOutput = simulationOutput)




# https://stats.stackexchange.com/questions/311556/help-interpreting-count-data-glmm-using-lme4-glmer-and-glmer-nb-negative-binom
# Optimize model fit using allFit, see http://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#troubleshooting

source(system.file("utils", "allFit.R", package="lme4"))
modelfit.all.m1.p <- allFit(m1.p)
ss1 <- summary(modelfit.all.m1.p)
ss1
```



# Testing `glmmTMB`

```{r}
p_load(glmmTMB)

data$SA3_NAME16 <- as_factor(data$SA3_NAME16)

m1 <- glmmTMB(Airbnb_visitor_nights ~ 1 + 
                (1 | SA3_NAME16),
              data = data, 
              offset = log(NVS_dest_nights_year),
              # family = genpois()
              family = nbinom2()
              # family = nbinom1()
)

m2 <- glmmTMB(Airbnb_visitor_nights ~ 1 + time + 
                (1 | SA3_NAME16),
              data = data, 
              offset = log(NVS_dest_nights_year),
              # family = genpois()
              family = nbinom2()
              # family = nbinom1()
)

m3 <- glmmTMB(Airbnb_visitor_nights ~ 1 + time + 
                (1 + time | SA3_NAME16),
              data = data, 
              offset = log(NVS_dest_nights_year),
              # family = genpois()
              family = nbinom2()
              # family = nbinom1()
)

summary(m3)
# coef(m3)
# ranef(m3)

data$m3 <- exp(predict(m3, data))

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m3),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")

ggplot(data, aes(x = Airbnb_visitor_nights)) +
  geom_point(aes(y = m3),
             alpha = 0.3) +
  geom_line(aes(y = Airbnb_visitor_nights), col = "darkorchid4") +
  facet_wrap(~date) +
  theme_minimal() + xlab("Airbnb_visitor_nights") + ylab("Predicted")
```
