==> R/98_LU_MB_clust.Rmd <==
---
title: "Land use clustering"
# subtitle: 
# author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    # css: custom.css
    toc: true
    toc_depth: 4
    toc_float: true
    number_sections: true
    theme: united 
    highlight: pygments 
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, 
                      fig.width=9, fig.height=7, dpi=300, out.width="900px", out.height="700px")

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

options(scipen=999)
set.seed(12345)

library(pacman) 
p_load(tidyverse, readr, readxl, lubridate, 
       scales, janitor, kableExtra, 
       sjmisc,
       sf, tmap)

tmap_mode("view")
```

```{r include=FALSE}
SA3 <- readRDS(file = "./data/geo/SA3_2016_AUST.rds")
# SA3_centr <- readRDS(file = "./data/geo/SA3_centr.rds")

```

# Land use

## Data

Mesh block categories aggregated to SA3 using area.

```{r}
MB_2016 <- read_csv("data/MB/MB_2016_join.zip") %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  filter(ste_name16 != "Other Territories") %>% 
  select(-objectid, -(sa1_main16:sa2_name16), -(sa4_code16:gcc_name16), 
         -(objectid_1:area_albers_sqkm), -(state:shape_area))

MB_2016_agg <- MB_2016 %>% 
  filter(mb_cat16 != "Water") %>% 
  group_by(sa3_code16, mb_cat16) %>% 
  summarize(area_cat16 = sum(areasqkm16)) %>% 
  ungroup() %>% 
  group_by(sa3_code16) %>% 
  mutate(area_tot = sum(area_cat16)) %>% 
  ungroup() %>% 
  mutate(area_cat16 = area_cat16 / area_tot) %>% 
  select(-area_tot)

MB_2016_agg_wide <- MB_2016_agg %>% 
  spread(mb_cat16, area_cat16) %>% 
  replace(is.na(.), 0) %>% 
  select(-sa3_code16) %>% 
  scale()
```

## Clustering 

### kmeans

```{r}
p_load(cluster, factoextra, NbClust, clValid)
```

```{r eval=FALSE, include=FALSE}
res.dist <- get_dist(MB_2016_agg_wide, stand = TRUE, method = "pearson")

fviz_dist(res.dist, 
          gradient = list(low = "#00AFBB", mid = "white", high = "#FC4E07"))
```

```{r eval=FALSE, include=FALSE}
fviz_nbclust(MB_2016_agg_wide, kmeans, k.max = 20, method = "silhouette")
fviz_nbclust(MB_2016_agg_wide, kmeans, k.max = 20, method = "wss")
fviz_nbclust(MB_2016_agg_wide, kmeans, k.max = 20, method = "gap_stat")

km.res <- kmeans(MB_2016_agg_wide, 9, nstart = 25)
```

```{r}
fviz_cluster(km.res, data = MB_2016_agg_wide, frame.type = "convex") +
  theme_minimal()
```

```{r}
MB_2016_agg_wide <- MB_2016_agg %>% 
  spread(mb_cat16, area_cat16) %>% 
  replace(is.na(.), 0)

MB_2016_agg_wide$cluster <- km.res$cluster

SA3 <- left_join(SA3, select(MB_2016_agg_wide, sa3_code16, cluster))

SA3 %>% 
  # filter(cluster == 9) %>% 
  tm_shape() +
  tm_polygons("cluster", palette = "Accent", style = "cat", n = 9,
              title = "Clusters", id = "SA3_NAME16")

temp <- MB_2016_agg_wide %>% 
  gather(Commercial:Transport, key = "key", value = "value")

temp %>% 
  ggplot(aes(x = key, value)) + 
  geom_boxplot(varwidth = TRUE) +
  facet_wrap(~cluster) +
  coord_flip() + 
  theme_light()

```


### other

```{r}
MB_2016_agg_wide <- MB_2016_agg %>% 
  spread(mb_cat16, area_cat16) %>% 
  replace(is.na(.), 0) %>% 
  select(-sa3_code16) %>% 
  scale()

pam.res <- pam(MB_2016_agg_wide, 9)

fviz_cluster(pam.res) +
  theme_minimal()
```


```{r}
# hc_method = "ward.D2"
res <- hcut(MB_2016_agg_wide, k = 9, stand = TRUE)

fviz_dend(res, rect = TRUE, cex = 0.5, k_colors = c("#00AFBB","#2E9FDF", "#E7B800", "#FC4E07"))
```

```{r}
res.nbclust <- NbClust(data = MB_2016_agg_wide, distance = "euclidean",
                       min.nc = 2, max.nc = 20, 
                       method = "complete", index ="all") 

factoextra::fviz_nbclust(res.nbclust) + 
  theme_minimal()
```

```{r}
res.hc <- eclust(MB_2016_agg_wide, "hclust", k = 9, graph = FALSE) 
fviz_silhouette(res.hc)
```

```{r}
intern <- clValid(MB_2016_agg_wide, nClust = 2:20, 
                  clMethods = c("hierarchical","kmeans","pam"),
                  validation = "internal")
# Summary
summary(intern)

p_unload(cluster, factoextra, NbClust, clValid)
```

### Mclust

```{r}
p_load(mclust)

BIC <- mclustBIC(MB_2016_agg_wide, G=1:20)
plot(BIC)
summary(BIC)

p_unload(mclust)
```

### VarSelLCM

```{r}
p_load(VarSelLCM)

# Cluster analysis without variable selection
res_without <- VarSelCluster(MB_2016_agg_wide, gvals = 1:20, vbleSelec = FALSE, crit.varsel = "BIC")

summary(res_without)
plot(res_without)

# Cluster analysis with variable selection (with parallelisation)
res_with <- VarSelCluster(MB_2016_agg_wide, gvals = 1:20, nbcores = 8, crit.varsel = "BIC")

summary(res_with)
print(res_with)

coef(res_with)

plot(res_with)

plot(x = res_with, y = "Commercial")
plot(x = res_with, y = "Education")
plot(x = res_with, y = "Hospital/Medical")
plot(x = res_with, y = "Industrial")
plot(x = res_with, y = "Parkland")
plot(x = res_with, y = "Primary Production")
plot(x = res_with, y = "Residential")
plot(x = res_with, y = "Transport")
plot(x = res_with, y = "Other")

plot(res_with, y = "Commercial", type = "cdf")


temp <- MB_2016_agg %>% 
  spread(mb_cat16, area_cat16) %>% 
  replace(is.na(.), 0)

temp$cluster <- fitted(res_with)

SA3$cluster <- NULL

SA3 <- left_join(SA3, select(temp, sa3_code16, cluster))

table(SA3$cluster)

tm_shape(SA3) +
  tm_polygons("cluster", palette="Accent", style="cat", n=5,
              title="Clusters", id="SA3_NAME16")

temp <- temp %>% 
  gather(Commercial:Transport, key = "key", value = "value")

temp %>% 
  ggplot(aes(x = key, value)) + 
  geom_boxplot(varwidth = TRUE) +
  facet_wrap(~cluster) +
  coord_flip() + 
  theme_light()

p_unload(VarSelLCM)

```
